Player character + nav mesh (basic)
I implemented a first person player and a basic nav mesh. Currently, the ball which the player is supposed to chase only moves towards one point, and very slowly. If at all possible, I would like for the ball to actively avoid the player and move quickly. If that's not possible, I thought the ball could move towards different wayponits (a lot of them) or move towards one object that is able to teleport across the map.